booking service class added
